// library-webapp/src/generated/library-api.ts
export interface components {
    schemas: {
        Book: {
            id: string;
            title: string;
            author: string;
            genre: string;
            description?: string;
            /** @enum {string} */
            status: "available" | "on-loan";
        };
        BookInput: {
            title: string;
            author: string;
            genre: string;
            description?: string;
        };
        Loan: {
            id: string;
            bookId: string;
            book?: components["schemas"]["Book"];
            memberId: string;
            /** Format: date */
            borrowedAt: string;
            /** Format: date */
            dueAt: string;
            /** Format: date */
            returnedAt?: string | null;
        };
        Error: {
            /** @description HTTP or application error code */
            code: number;
            /** @description short human-readable label */
            message: string;
            /** @description detailed explanation */
            description?: string;
            /** @description URI to documentation */
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}